docs: Improve troubleshooting documentation
Due to the temporarily doubled ESP space usage, it is now easier to run
into the out of space issue (once). Use this opportunity to document how
to proceed in this case.

Furthermore, recovery in case of ESP corruption is now slightly more
involved, because not all files are rewritten all the time. Document the
full recovery instructions.